+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 - Pick point from point cloud to display (x,y,x) values in terminal added.
 - Origin axis frame added to pointcloud visualisation.
 - Live calibration function added.
+- Example file added for live calibration.
 
 [0.0.2] - 2025-03-06
 --------------------