+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Added
 - Live calibration function added.
 - Example file added for live calibration.
 - Added stereo camera detection by name function.
+- Added rectification of coordinates from camera parameters.
 
 [0.0.2] - 2025-03-06
 --------------------