+++ CHANGELOG.rst
@@ -9,8 +9,8 @@ and this project adheres to `Semantic Versioning`_.
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0/
 .. _Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]
-------------
+[0.0.3] - 2025-08-24
+--------------------
 
 Added
 ^^^^^